# openapi.yaml
openapi: 3.1.0
info:
  title: Audicle
  version: 0.57.0
paths:
...
  /api/v1/settings:
    get:
      tags:
      - settings
      summary: Get Settings Overrides
      operationId: get_settings_overrides_api_v1_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
      security:
      - {}
      - AdminSession: []
    put:
      tags:
      - settings
      summary: Put Settings Overrides
      operationId: put_settings_overrides_api_v1_settings_put
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Payload
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
  /api/v1/settings/ocr/languages:
    get:
      tags:
      - settings
      summary: Get Ocr Languages
      description: Languages the shipped OCR model packs cover; drives the Settings
        dropdown.
      operationId: get_ocr_languages_api_v1_settings_ocr_languages_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrLanguagesResponse'
      security:
      - {}
      - AdminSession: []
...
components:
  schemas:
...
    OcrLanguagesResponse:
      properties:
        languages:
          items:
            type: string
          type: array
          title: Languages
        default:
          type: string
          title: Default
      additionalProperties: false
      type: object
      required:
      - languages
      - default
      title: OcrLanguagesResponse
...
    SettingsResponse:
      properties:
        allowlist:
          items:
            type: string
          type: array
          title: Allowlist
        values:
          additionalProperties: true
          type: object
          title: Values
        defaults:
          additionalProperties: true
          type: object
          title: Defaults
        feed_url:
          type: string
          title: Feed Url
      additionalProperties: false
      type: object
      required:
      - allowlist
      - values
      - defaults
      - feed_url
      title: SettingsResponse
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - status
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: audicle_session
      description: Required for the admin API when a password is configured.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the admin session for unsafe methods.
    FeedKey:
      type: apiKey
      in: query
      name: key
      description: Required for protected feeds and non-artwork media.